Got rid of unique pointers, added setters and overload for setBandFrequencies, + more
filters is now a vector of dsp::IIR::filter<float> objects instead of unique pointers to these objects, which makes the code more understandable for us. sampleRate and qFactor are now private variables of the FilterBank that can be changed with setter functions. setBandFrequencies() now does not need parameters for centerFrequencies, as they will be calculated using the frequency range, but there is an overload in case we want to pass in custom center frequencies anyway. Added lots of comments.